WIP type refactoring (major readability boost!)
