+++ crates/crypto/src/keys/keyring/mod.rs
@@ -1,4 +1,4 @@
-use crate::{primitives::SecretKeyString, Protected, Result};
+use crate::{primitives::types::SecretKeyString, Protected, Result};
 
 #[cfg(target_os = "linux")]
 pub mod linux;
