+++ crates/crypto/src/protected.rs
@@ -40,17 +40,6 @@ where
 	data: T,
 }
 
-impl<T> std::ops::Deref for Protected<T>
-where
-	T: Zeroize,
-{
-	type Target = T;
-
-	fn deref(&self) -> &Self::Target {
-		&self.data
-	}
-}
-
 impl<T> Protected<T>
 where
 	T: Zeroize,
